ingest/requirements.txt
sentence-transformers
feedparser
newspaper3k
requests
tqdm
torch
